Rename func_test package to e2e
