+++ e2e/peer_ops_test.go
@@ -1,4 +1,4 @@
-package func_test
+package e2e
 
 import (
 	"fmt"
